tileMap avec un TilePane

+++ src/main/java/oumaima_nezha_mehdy/zelda/Univers/Terre.java
@@ -2,7 +2,7 @@ package oumaima_nezha_mehdy.zelda.Univers;
 
 
 public class Terre extends Bloc{
-    public Terre(int x,int y) {
+    public Terre(double x,double y) {
         super(true,x,y);
     }
